Moved remove account to it's own class to be able to reuse it in admin class
