+++ src/main/account/CreateAccount.java
@@ -2,5 +2,4 @@ package main.account;
 
 public class CreateAccount {
     private static int accountIdCounter = -1;
-
 }
